kindle-crossword/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.8.0
pydantic-settings==2.4.0
Pillow==10.4.0
numpy>=1.26.0
opencv-python-headless==4.9.0.80
img2pdf==0.5.1
aiofiles==24.1.0
